# SBT-Documentation/swagger.yaml
openapi: 3.0.3
info:
  title: Sales By Twilight API
  description: |
    Professional REST API for a grocery store e-commerce platform.
    
    ## Features
    - Full CRUD for Categories, Products, Customers, and Orders
    - Consistent JSON response format
    - Comprehensive filtering and sorting
    - Stock management with automatic inventory tracking
    - Order lifecycle with status workflow
    
    ## Response Format
    All responses follow a consistent structure:
    - Success: `{success: true, data: ..., message: "..."}`
    - Error: `{success: false, error: {code: "...", message: "..."}}`
    
  version: 1.0.0
  contact:
    name: Nkem
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:5001
    description: Development server
...
paths:
...
  /categories:
    get:
      tags: [Categories]
      summary: List all categories
      parameters:
        - name: name
          in: query
          schema:
            type: string
          description: Filter by name (partial match)
        - name: sort
          in: query
          schema:
            type: string
            enum: [id, name, created_at, updated_at]
          description: Sort field
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
          description: Sort order
        - name: limit
          in: query
          schema:
            type: integer
          description: Maximum results
        - name: offset
          in: query
          schema:
            type: integer
          description: Skip results
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessListResponse'
    post:
      tags: [Categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /categories/{id}:
    get:
      tags: [Categories]
      summary: Get a category
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Categories]
      summary: Update a category
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
      responses:
        '200':
          description: Category updated
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Categories]
      summary: Delete a category
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Category deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ============ PRODUCTS ============
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
    SuccessListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    CategoryCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
...